Lots of bug fixes.
Behavior:
sceneSummary will not be generated if not needed.
Do not load the cache if it is not needed. Slight conflict with this.
Improvements:
sceneSummary output double newlines reduced by one.
Added colon : between file and lines numbers when generating sceneSummary.
Always print out model and version information.
Removed old code.
Fixed:
speakerName in settings{} was not getting reset.
contextHistory was not formatted correctly for gemma-2 models.
Update mainSpreadsheet even if all entries were found in the cache.